Added validations file
