--- client/src/utils/getHasMaterials.js
@@ -1,3 +1,0 @@
-export default function getHasMaterials(materials, inventory) {
-  return materials.every((material) => inventory.find(item => item.type === material.type));
-}
